Add initial cargo-fuzz usage.
I used this to find a panic in the parser already.  Brief HOWTO:

Install:
    cargo install cargo-fuzz

Pick a fuzzer and run it:
    cargo fuzz run parse_integral_lit

Now that the fuzzer spit out a result (bytes to the console, and a
file on the filesystem), construct a minimal testcase:

    cargo fuzz tmin parse_integral_lit \
      fuzz/artifacts/parse_integral_lit/crash-whatever

Now you have a new file at
fuzz/artifacts/parse_integral_lit/minimized-from-whatever, turn that
into a testcase in cargo and fix it!

Fixes #1