Fix and optimize parsers that must fail in branch parsers. #85, #101
Sometimes, parsers fail and failure is okay. For instance, in
`opt(parser)`, if `parser` fails, the `opt` parser will turn the error
into a continuation of the parent parser. That is wanted to authorize
having more interesting grammars and is necessary.

However, sometimes, we might want to broadcast a kind of error that is
not recoverable from such parent parsers. For instance, in
`opt(pair(parser_a, parser_b))`, if `parser_a` fails, it means that the
whole parent parser can just treat the whole `opt` has errored. Now, if
`parser_a` succeeds but `parser_b` fails, it means two things:

- Either this is okay if the parent parser might want to parse with
  `parser_a` again.
- Or this is wrong, especially if `parser_a` identifies the whole
option.

For instance, parsing an optional assignement operation (e.g. `=
something`) **must** fail if we have parsed `=` but not `something`.

This very commit fixes that situation for **all** parsers. The immediate
effect of this fix is:

- We now see GLSL errors as real errors instead of removing branches and
returning false-positives.
- Some parsers shortcut _way_ more quickly now, improving performances.

Fix both #85 and #101.

Courtesy for @Eijebong, mandatory commit gif.

![](https://phaazon.net/media/uploads/eyebrows.gif)